move sys log to multislogger package, use default logger on error instead of exit
